+++ code/ui.shell.doc/README.md
@@ -1,0 +1,24 @@
+![Module](https://img.shields.io/badge/%40platform-ui.shell.doc-%23EA4E7E.svg)
+[![License: MIT](https://img.shields.io/badge/license-MIT-blue.svg)](https://opensource.org/licenses/MIT)
+[![NPM](https://img.shields.io/npm/v/@platform/ui.shell.doc.svg?colorB=blue&style=flat)](https://www.npmjs.com/package/@platform/ui.shell.doc)
+![banner](https://user-images.githubusercontent.com/185555/64071664-f4933f80-ccd2-11e9-87cb-8713ad2fb868.png)
+
+Document components that operate within [ui.shell](../ui.shell)
+
+<p>&nbsp;<p>
+
+## Live Demo:
+https://shell.platform.uiharness.com  
+See corresponding test module: [`ui.shell.test`](../ui.shell.test)
+
+<p>&nbsp;<p>
+
+
+## Setup
+
+    yarn add @platform/ui.shell.doc
+
+
+
+<p>&nbsp;<p>
+<p>&nbsp;<p>
