+++ code/ui.shell.doc/package.json
@@ -1,0 +1,28 @@
+{
+  "name": "@platform/ui.shell.doc",
+  "version": "0.0.0",
+  "description": "Document components that operate within ui.shell.",
+  "main": "lib/index",
+  "types": "lib/index.d.ts",
+  "scripts": {
+    "test": "ts test",
+    "tdd": "ts test --watch",
+    "lint": "ts lint",
+    "build": "ts build",
+    "prepare": "ts prepare"
+  },
+  "dependencies": {
+    "@platform/libs": "0.0.12",
+    "@platform/react": "0.3.20",
+    "@platform/ui.shell.types": "0.0.0"
+  },
+  "devDependencies": {
+    "@platform/test": "0.0.50",
+    "@platform/ts": "3.4.17"
+  },
+  "files": [
+    "lib",
+    "!*.TEST.{js, mjs, d.ts}"
+  ],
+  "license": "MIT"
+}
