+++ code/ui.shell.doc/src/common/constants.ts
@@ -1,0 +1,1 @@
+export const TEMP = 123;
