+++ code/ui.shell.doc/src/common/index.ts
@@ -1,0 +1,5 @@
+import * as t from './types';
+import * as constants from './constants';
+
+export { constants, t };
+export * from './libs';
