+++ code/ui.shell.doc/src/common/libs.ts
@@ -1,0 +1,2 @@
+import { uniq } from 'ramda';
+export const R = { uniq };
