+++ code/ui.shell.doc/src/common/types.ts
@@ -1,0 +1,1 @@
+export * from '../types';
