+++ code/ui.shell.doc/tsconfig.json
@@ -1,0 +1,7 @@
+{
+  "extends": "@platform/ts/tsconfig.json",
+  "compilerOptions": {
+    "rootDir": "src",
+    "outDir": "lib"
+  }
+}
