+++ code/ui.shell.types/README.md
@@ -1,0 +1,24 @@
+![Module](https://img.shields.io/badge/%40platform-ui.shell.types-%23EA4E7E.svg)
+[![License: MIT](https://img.shields.io/badge/license-MIT-blue.svg)](https://opensource.org/licenses/MIT)
+[![NPM](https://img.shields.io/npm/v/@platform/ui.shell.types.svg?colorB=blue&style=flat)](https://www.npmjs.com/package/@platform/ui.shell.types)
+![banner](https://user-images.githubusercontent.com/185555/64071664-f4933f80-ccd2-11e9-87cb-8713ad2fb868.png)
+
+Shared types used by [ui.shell](../ui.shell) components.
+
+<p>&nbsp;<p>
+
+## Live Demo:
+https://shell.platform.uiharness.com  
+See corresponding test module: [`ui.shell.test`](../ui.shell.test)
+
+<p>&nbsp;<p>
+
+
+## Setup
+
+    yarn add @platform/ui.shell.types
+
+
+
+<p>&nbsp;<p>
+<p>&nbsp;<p>
