+++ code/ui.shell.types/package.json
@@ -1,0 +1,27 @@
+{
+  "name": "@platform/ui.shell.types",
+  "version": "0.0.0",
+  "description": "Shared types used by ui.shell components.",
+  "main": "lib/index",
+  "types": "lib/index.d.ts",
+  "scripts": {
+    "test": "ts test",
+    "tdd": "ts test --watch",
+    "lint": "ts lint",
+    "build": "ts build",
+    "prepare": "ts prepare"
+  },
+  "dependencies": {
+    "@platform/libs": "0.0.12",
+    "@platform/ui.tree": "0.3.0"
+  },
+  "devDependencies": {
+    "@platform/test": "0.0.50",
+    "@platform/ts": "3.4.17"
+  },
+  "files": [
+    "lib",
+    "!*.TEST.{js, mjs, d.ts}"
+  ],
+  "license": "MIT"
+}
