+++ code/ui.shell.types/tslint.json
@@ -1,0 +1,4 @@
+{
+  "extends": "@platform/ts/tslint.json",
+  "rules": {}
+}
