+++ code/ui.shell/package.json
@@ -13,9 +13,11 @@
     "prepare": "ts prepare"
   },
   "dependencies": {
+    "@platform/log": "0.0.140",
     "@platform/react": "0.3.20",
     "@platform/ui.icon": "0.0.84",
     "@platform/ui.loader": "0.1.7",
+    "@platform/ui.shell.types": "0.0.0",
     "@platform/ui.tree": "0.3.0",
     "@platform/util.value": "0.4.11"
   },
