+++ code/ui.shell/src/api/index.ts
@@ -1,2 +1,1 @@
-export * from './types';
 export * from './Shell';
