+++ code/ui.shell/src/common/types.ts
@@ -1,7 +1,5 @@
 export * from '../types';
-export * from '../state/types';
 
 export * from '@platform/ui.loader/lib/types';
 export * from '@platform/ui.tree/lib/types';
-
 export { IObservableProps } from '@platform/util.value/lib/types';
