+++ code/ui.shell/src/components/common.ts
@@ -1,6 +1,6 @@
 import * as state from '../state';
 
 export { state };
-export { Context, createProvider } from '../context';
+export { Context, createProvider } from '../Context';
 export { Shell } from '../api';
 export * from '../common';
