--- code/ui.shell/src/components/types.ts
@@ -1,3 +1,0 @@
-import * as t from '../common/types';
-
-export type ShellTheme = t.LoaderTheme;
