--- code/ui.shell/src/context/index.ts
@@ -1,2 +1,0 @@
-export * from './Context';
-export * from './types';
