--- code/ui.shell/src/context/types.ts
@@ -1,8 +1,0 @@
-import * as t from '../common/types';
-
-export type IShellContext = {
-  // loader: t.ILoader;
-  shell: t.IShell;
-  // splash: t.ISplash;
-  // theme: t.ShellTheme;
-};
