+++ code/ui.shell/src/state/index.ts
@@ -1,4 +1,2 @@
 import * as shell from './shell';
-
 export { shell };
-export * from './types';
